Move geography endpoint definition into new file
